voice/requirements.txt
sounddevice==0.4.6
numpy==1.26.4
openwakeword==0.6.0
onnxruntime==1.17.3
vosk==0.3.45
websockets==12.0
pyttsx3==2.90
pyyaml==6.0.1
